requirements.txt
# Package and package dependencies
-e .

# Development dependencies
setuptools>=68.2.2,<68.3
wheel>=0.41.2,<0.42
twine>=4.0.2,<4.1
coverage>=7.3.2,<7.4

# Linters and formatters
pre-commit>=3.5.0,<3.6

# TestApp dependencies
django>=3.2,<4.3
jsonschema>=4.19.1,<4.20
djangorestframework>=3.14.0,<4
django-generic-contact>=1.0.0,<1.1
